+++ frontend/src/components/screens/create_ticket/rs/attachments_modal/page.tsx
@@ -1,7 +1,6 @@
 
 "use client"
-import React from 'react'
-import { Tldraw } from 'tldraw';
+import { Button } from '@/components/ui/button';
 import {
     Dialog,
     DialogContent,
@@ -9,7 +8,6 @@ import {
     DialogHeader,
     DialogTitle
 } from "@/components/ui/dialog";
-import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 
 type TAttachmentsModal = {