+++ frontend/src/components/screens/departments/dtv_ha/technicians/tasks_update/page.tsx
@@ -39,7 +39,6 @@ import repairshopr_statuses from "@/lib/repairshopr_status";
 import { cn } from "@/lib/utils";
 import warranties from "@/lib/warranties";
 
-
 type TTasksUpdate = {
     assessment_date: string;
     service_order_noProp: string | number | undefined