+++ frontend/src/components/screens/departments/hhp/technicians/qc/page.tsx
@@ -1,16 +1,16 @@
 "use client"
 
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { datetimestamp } from '@/lib/date_formats';
-import React, { ChangeEvent } from "react";
 import {
     Accordion,
     AccordionContent,
     AccordionItem,
     AccordionTrigger,
 } from "@/components/ui/accordion";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Textarea } from "@/components/ui/textarea";
+import { datetimestamp } from '@/lib/date_formats';
+import React, { ChangeEvent } from "react";
 type TQC = {
     qc_fail_reasonProp: string
     qc_completeProp: string