+++ frontend/src/components/screens/driver_app/create_checklist/multistep_form/section_nine.tsx
@@ -1,5 +1,5 @@
 import { Textarea } from '@/components/ui/textarea';
-import React, { ChangeEvent } from 'react';
+import React from 'react';
 import FormWrapper from './wrapper';
 
 interface ISectionNine {