+++ server/controllers/department/hhp/technicians/add_hhp_task.js
@@ -79,6 +79,7 @@ const AddHHPTask = async (req, res) => {
             });
         }
     } catch (error) {
+        console.log(error);
         // Handle validation or other errors
         const errors = {};
         if (error.inner) {